refactor: extract onboarding cli module
